requirements.txt
Django==2.2.19
djangorestframework==3.13.1
djoser==2.1.0
python-dotenv==0.20.0
webcolors==1.11.1
psycopg2-binary==2.9.3
pillow==9.1.0
django-filter==21.1